backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Pydantic
pydantic[email]==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Twilio
twilio==8.10.0

# OpenAI
openai==1.12.0

# Email
aiosmtplib==3.0.1

# HTTP client
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==24.1.0
isort==5.13.2
